requirements.txt
gradio
ultralytics
opencv-python-headless
numpy
pillow
torch
